+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+v1.0.2 (2019-05-05)
+* Fixed ability to use properties in programs
+* Added GV8 Locked state
+* Added GV9 Sun roof percent
+* Added GV10 Odometer
+* State of Charge % has been moved to ST 
+* Online state has been moved to GV18
+
 v1.0.1 (2019-04-20)
 * Added support for Polyglot Cloud.